requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
azure-ai-projects==1.0.0
azure-ai-agents==1.1.0
azure-identity==1.24.0
python-dotenv==1.1.1